separate main excecution from other class
